requirements.txt
open3d
numpy
pillow
opencv-python
torch
albumentations